pyproject.toml
[build-system]
requires = [
  "hatch-vcs>=0.3",
  "hatchling>=1.12.2",
]
build-backend = "hatchling.build"

[project]
name = "monailisa"
authors = [
    {name = "Nathan Hoche", email = "hochenat@gmail.com"},
]
description = "Genetic Algorithm to recreate image"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
url.Homepage = "https://github.com/nathan-hoche/MonaLisa"
url.Source = "https://github.com/nathan-hoche/MonaLisa"
url.Tracker = "https://github.com/nathan-hoche/MonaLisa/issues"
classifiers = [
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Multimedia :: Graphics",
    "Topic :: Utilities"
]
dynamic = ["version"]
dependencies = [
    "scikit-image",
    "pillow",
    "numpy",
]

[project.optional-dependencies]
dev = ["black", "isort", "pip-tools", "build", "twine", "pre-commit", "commitizen"]

[project.scripts]
monailisa = "monailisa.__init__:main"

[options.packages.find]
where = "monailisa"
exclude = "tests*"

[tool.hatch]
version.source = "vcs"
